+++ digital_signature/__manifest__.py
@@ -1,9 +1,8 @@
 # -*- coding: utf-8 -*-
 #############################################################################
-#
 #    Cybrosys Technologies Pvt. Ltd.
 #
-#    Copyright (C) 2022-TODAY Cybrosys Technologies(<https://www.cybrosys.com>)
+#    Copyright (C) 2023-TODAY Cybrosys Technologies(<https://www.cybrosys.com>)
 #    Author: Cybrosys Techno Solutions(<https://www.cybrosys.com>)
 #
 #    You can modify it under the terms of the GNU LESSER
@@ -19,28 +18,37 @@
 #    If not, see <http://www.gnu.org/licenses/>.
 #
 #############################################################################
-
 {
     'name': 'Digital Signature In Purchase Order, Invoice, Inventory',
-    'summary': 'Digital Signature in Purchase Order, Invoice, Inventory V15',
-    'version': '16.0.1.0.0',
-    'category': 'Tools',
-    'description': """Digital Signature in Purchase Order, Invoice, Inventory V15""",
+    'version': '16.0.2.1.0',
+    'category': 'Extra Tools',
+    'summary': """This module provide feature to add the digital signature and 
+    company stamp.""",
+    'description': """Module helps to add the digital signature and company
+     stamp to the report of the purchase order, invoice, and inventory.""",
     'author': 'Cybrosys Techno Solutions',
     'company': 'Cybrosys Techno Solutions',
     'maintainer': 'Cybrosys Techno Solutions',
+    'website': 'https://www.cybrosys.com',
     'depends': ['purchase', 'stock', 'account'],
     'data': [
-        'views/res_config_settings.xml',
-        'views/purchase_order.xml',
-        'views/inventory.xml',
-        'views/invoice.xml',
-        'views/purchase_report_inherit.xml',
-        'views/stock_picking_report.xml',
-        'views/invoice_report.xml',
+        'views/purchase_order_views.xml',
+        'views/account_move_views.xml',
+        'views/purchase_order_templates.xml',
+        'views/stock_picking_views.xml',
+        'views/account_move_templates.xml',
+        'views/res_company_views.xml',
+        'views/stock_picking_templates.xml',
+        'views/res_config_settings_views.xml',
     ],
-    'installable': True,
-    'application': False,
+    'assets': {
+        'web.report_assets_common': [
+            'digital_signature/static/src/css/signature_stamp.css',
+        ],
+    },
     'images': ['static/description/banner.png'],
     'license': 'LGPL-3',
+    'installable': True,
+    'auto_install': False,
+    'application': False,
 }