+++ digital_signature/models/res_company.py
@@ -1,0 +1,37 @@
+# -*- coding: utf-8 -*-
+#############################################################################
+#    Cybrosys Technologies Pvt. Ltd.
+#
+#    Copyright (C) 2023-TODAY Cybrosys Technologies(<https://www.cybrosys.com>)
+#    Author: Cybrosys Techno Solutions(<https://www.cybrosys.com>)
+#
+#    You can modify it under the terms of the GNU LESSER
+#    GENERAL PUBLIC LICENSE (LGPL v3), Version 3.
+#
+#    This program is distributed in the hope that it will be useful,
+#    but WITHOUT ANY WARRANTY; without even the implied warranty of
+#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+#    GNU LESSER GENERAL PUBLIC LICENSE (LGPL v3) for more details.
+#
+#    You should have received a copy of the GNU LESSER GENERAL PUBLIC LICENSE
+#    (LGPL v3) along with this program.
+#    If not, see <http://www.gnu.org/licenses/>.
+#
+#############################################################################
+from odoo import fields, models
+
+
+class ResCompany(models.Model):
+    """Inherited the res company for setting the company stamp"""
+
+    _inherit = 'res.company'
+    _description = 'Company'
+
+    stamp = fields.Binary(string="Stamp", help="Company stamp")
+    position = fields.Selection(selection=[
+        ('left', 'Left'),
+        ('right', 'Right'),
+        ('center', 'Center'),
+    ], string='Position', tracking=True, default="left",
+        help="The position is used  to position the  signature and company "
+             "stamp in reports")