+++ digital_signature/doc/RELEASE_NOTES.md
@@ -1,6 +1,11 @@
 ## Module <digital_signature>
 
-#### 19.10.2023
+#### 01.07.2022
 #### Version 16.0.1.1.0
 #### ADD
-- Initial commit for Odoo 16 Digital Signature
+- Initial commit for Digital Signature In Purchase Order, Invoice, Inventory
+
+#### 26.10.2023
+#### Version 16.0.2.1.0
+#### UPDT
+- Added the company stamp feature in  Purchase Order, Invoice and Inventory